+++ CotEditor/Sources/CESplitViewController.m
@@ -50,6 +50,17 @@
 
 
 // ------------------------------------------------------
+/// clean up
+- (void)dealloc
+// ------------------------------------------------------
+{
+    // Need to set nil to NSSPlitView's delegate manually since it is not weak but just assign,
+    //     and may crash when closing split fullscreen window on El Capitan beta 5 (2015-07)
+    [[self splitView] setDelegate:nil];
+}
+
+
+// ------------------------------------------------------
 /// 自身の view として NSSplitView を返す (NSSplitViewController のメソッド)
 - (nonnull NSSplitView *)splitView
 // ------------------------------------------------------